Add feature to allow any model to use normalized data
