+++ model/base_model.py
@@ -8,6 +8,9 @@ Last updated: MB 29/08/2020 - created module.
 import pandas as pd
 import matplotlib.pyplot as plt
 
+# import local modules.
+from utils import data_loader
+
 """
 Base class that all models inherit from.
 """
@@ -15,11 +18,22 @@ class BaseModel:
     """
     store dataset. data is a dictionary.
     """
-    def __init__(self, data):
+    def __init__(self, data, normalize=False):
         print(">> initialising model...")
-        self.train_x = data['train_x']
+
+        # if we are normalizing data, save the normalized x value.
+        if normalize is True:
+            self.normalization_params = data_loader.get_normalization_params(data['train_x'])
+            self.train_x = self.normalize_x(data['train_x'])
+            self.test_x = self.normalize_x(data['test_x'])
+
+        # if we are not normalizing data, use regular x values.
+        else:
+            self.train_x = data['train_x']
+            self.test_x = data['test_x']
+
+        # save the y values and other attributes.
         self.train_y = data['train_y']
-        self.test_x = data['test_x']
         self.test_y = data['test_y']
         self.test_predictions = pd.Series()    # placeholder for 'test' function.
         self.is_trained = False
@@ -91,3 +105,28 @@ class BaseModel:
 
         # calculate the correlation.
         print("correlation: %.2f" % self.test_y.corr(self.test_predictions))
+
+    """
+    Convert a pandas dataframe of values into normalized values based on the
+    normalized params attribute. x_values is a pandas dataframe.
+    """
+    def normalize_x(self, x_values):
+        # throw an error if this model was not setup to use normalized values.
+        if not self.normalization_params:
+            raise Exception("This model was not setup to use normalized values.")
+            return
+
+        # copy the dataframe.
+        normalized_values = pd.DataFrame()
+
+        # iterate over each column and normalize.
+        for column in x_values.columns:
+            # retrieve normalization parameters.
+            mean = self.normalization_params[column]['mean']
+            std = self.normalization_params[column]['std']
+
+            # save the normalized column.
+            normalized_values[column] = (x_values[column] - mean) / std
+
+        # return the normalized dataframe.
+        return normalized_values
