+++ model/least_squares.py
@@ -18,9 +18,9 @@ class LeastSquares(BaseModel):
     """
     initialise class instance.
     """
-    def __init__(self, data):
+    def __init__(self, data, normalize=False):
         # call parent function.
-        BaseModel.__init__(self, data)
+        BaseModel.__init__(self, data, normalize=normalize)
 
         # placeholders specific to this class.
         self.model = None
